+++ functions/rmp.php
@@ -5,7 +5,11 @@ if(empty($_GET['id'])){
 	die("Modpack not specified.");
 }
 if(!$_SESSION['user']||$_SESSION['user']=="") {
-	die("You need to be logged in!");
+	die("Unauthorized require or login session has expired!");
+}
+if(substr($_SESSION['perms'],0,1)!=="1") {
+	echo 'Insufficient permission!';
+	exit();
 }
 mysqli_query($conn, "DELETE FROM `builds` WHERE `modpack` = '".$_GET['id']."'");
 mysqli_query($conn, "DELETE FROM `modpacks` WHERE `id` = '".$_GET['id']."'");
